Use the detailed_merge_status of a mr to refine the action when updating MRs
This tries to be a little smart about whether or not it makes sense to attempt an actual merge of an MR.
Depending on the detailed merge status, we can know that a merge will not succeed.
In such cases, the merge request will remain untouched and the user will be informed.