Use `gh` instead of a dedicated action for working with releases
Creation of new draft releases and uploading release artifacts is now
done using the GitHub CLI instead of some action.
The previous, official action is deprecated and instead of choosing an
action by someone from the internet, it feels better to use the official
CLI tool that's installed on the runners already.

Closes #16